[up] update cli structure
